jED

Documentation for jED.

jED.AIM

jED.Basis

jED.Eigenspace

jED.Eigenspace

jED.CDag_sign

jED.C_sign

jED.N_el

jED.S

jED._H_CDagC

jED._H_nn

jED._generate_blocks!

jED.calc_E

jED.calc_Hamiltonian

jED.calc_Z

jED.operator_ni

jED.overlap

jED.overlap_cdagger_c

jED.overlap_ni_nj

jED.AIM — Type

AIM <: Model
AIM(ϵₖ::Vector{Float64}, Vₖ::Vector{T}, μ::Float64, U::Float64)

Type for the Anderson impurity model. A model can be used to construct a Eigenspace given a set of Basis.

Can be created from

bath onsite energies $\epsilon_k$,

bath-impurity hopping $V_k$

impurity chemical potential $\mu$

Coulomb interaction strength on the impurity $U$

Example

``` julia> ϵₖ = [0.5, -5.0] julia> Vₖ = [1.0, 1.0] julia> U = 1.0 julia> μ = 0.5 julia> model = AIM(ϵₖ, Vₖ, μ, U) AIM{3, Float64}([ -0.5 1.0 1.0; 1.0 0.5 0.0; 1.0 0.0 -5.0], [1.0 0.0 0.0; 0.0 0.0 0.0; 0.0 0.0 0.0], [0.0 0.0 0.0; 0.0 0.0 0.0; 0.0 0.0 0.0])

jED.Basis — Type

Basis

A state is represented as a BitVetor of length 64. The first 32 bit (ordering follows Julia Vector indexing) correspond to spin up for each site (index 1 corresponds to the first site, 2 to the second, etc.) and bits 33 to 64 to spin down, for site 1 to 32. Each bit is set to 1 if it is occupied by an electron of the given spin. In principle, the number of electron flavors and orbitals is free, however, for now it is fixed to 2 and 1.

A list of states should be generated using States(NSites) with NSites being the number of sites in the model.

Fields

NFlavors : Integer, number of flavors (TODOhardcoded to 2 for now!)

NSites : Integer, number of sites

states : Vector{State}, list of Fock states. The are sorted according to good quantum numbers. blocklist contains start and size of blocks.

blocklist : Vector{Tuple{Int,Int}}, Vector of tuples. Each entry encodes a block with the first element being the start and the second the column size of the block

jED.Eigenspace — Type

Eigenspace

Containes Eigenvalues and Eigenvectors grouped into blocks with are each ordered by magnitude of Eigenvalues. The blocks are part of the Basis.

Fields

evals : Eigenvalues

evecs : Eigenvectors

blocklist : List of start indices and lengths of blocks.

E0 : smallest Eigenvalue

jED.Eigenspace — Method

Constructs Eigenspace for Model over given Basis by diagonalizing the Hamiltonian (see also calc_Hamiltonian) for each block.

jED.CDag_sign — Method

CDag_sign(state,i)

Sign when annihilating an electron at position i in state. Returns: -1/1/0 for Uneven/Even permutations and 0 when the state is not occupied.

julia> CDag_sign(Fockstate{6}(Bool[1,0,1,1,0,0]),5)
-1
julia> CDag_sign(Fockstate{6}(Bool[1,0,1,1,0,0]),3)
0
julia> CDag_sign(Fockstate{7}(Bool[1,0,1,0,1,0,0]),4)
1

jED.C_sign — Method

C_sign(state,i)

Sign when creating an electron at position i in state. Returns: -1/1/0 for Uneven/Even permutations and 0 when the state is not occupied.

julia> C_sign(Fockstate{6}(Bool[1,0,1,1,0,0]),3)
-1
julia> C_sign(Fockstate{6}(Bool[1,0,1,1,0,0]),2)
0
julia> C_sign(Fockstate{6}(Bool[1,0,1,1,0,0]),4)

jED.N_el — Method

N_el(s::AbstractVector)::Int

Total electron number of state.

jED.S — Method

S(s::AbstractVector)::Int

Total spin of state. #TODO: only implemented for flavor=2

jED._H_CDagC — Method

_H_CDag_C(istate, jstate, tMatrix)

Returns the hopping contribution for states $\sum_{i,j} \langle i | T | j \rangle$, with T being the hopping matrix tmatrix and the states i and j given by istate and jstate.

jED._H_nn — Method

_H_nn(istate, jstate, tMatrix)

Returns the hopping contribution for states $\sum_{i,j} \langle i | T | j \rangle$, with T being the hopping matrix tmatrix and the states i and j given by istate and jstate.

jED._generate_blocks! — Method

_generate_blocks!(states::Vector{Fockstate})

Sort state list and generate list of blocks.

jED.calc_E — Method

calc_E(eigenspace, β)

Calculates the total energy from eigenspace and the inverse temperature β.

jED.calc_Hamiltonian — Method

calc_Hamiltonian(model::Model, basis::Basis)

Calculates the Hamiltonian for a given

model, see for example AIM)) in a

basis, see Basis

jED.calc_Z — Method

calc_Z(eigenspace, β)

Calculates the partition function from eigenspace and the inverse temperature β.

jED.operator_ni — Method

operator_ni(state::Fockstate, i::Int)

Calculate n_i |ket⟩, i.e. i is the index for the number operator

Returns: (eigenval,newState)::Tuple{Int, Fockstate} with eigenval 1 (state i occupied) or 0 (state i unoccupied) and newState. In case of this operator, newState === state. The density overlap of 2 states can be calculated efficiently using overlap_ni_nj.

Example

``` julia> jED.operator_ni(jED.SVector{8}(Bool[1,1,1,1,0,1,0,1]), 2); (1, Bool[1, 1, 1, 1, 0, 1, 0, 1])

jED.overlap — Method

overlap(bra::Fockstate, ket::Fockstate)

Calculates ⟨bra|ket⟩. This is useful when operators have been applied to one state and no efficient direct implementation for an overlap involving this operator is available.

Returns: True/False

jED.overlap_cdagger_c — Method

overlap_cdagger_c(bra::Fockstate, i::Int, ket::Fockstate, j::Int)::Int

Calculate ⟨bra| c^†i cj |ket⟩, i.e. i is the index for the creation operator and j the index for the annihilation operator. Internally, we check, that both states have exactly two or no occupation difference. In both cases the product of C_sign and CDag_sign is returned, otherwise 0. Note, that this allows spin flips!

Returns: -1/0/1

Example

julia> t1 = jED.SVector{8}(Bool[1,1,1,1,0,1,0,1]);
julia> t2 = jED.SVector{8}(Bool[1,1,0,1,1,1,0,1]);
julia> jED.overlap_cdagger_c(t1,t2,3,3)
0
julia> jED.overlap_cdagger_c(t1,t1,2,2)
1

jED.overlap_ni_nj — Method

overlap_ni_nj(bra::Fockstate, ket::Fockstate, i::Int, j::Int)

Calculate ⟨bra| n^†i nj |ket⟩

Returns: True/False (converts to 1/0)