jED

Documentation for jED.

jED.States

jED.CDag_sign

jED.C_sign

jED.N_el

jED.S

jED.overlap_cdagger_c

jED.States — Type

States

A state is represented as a BitVetor of length 64. The first 32 bit (ordering follows Julia Vector indexing) correspond to spin up for each site (index 1 corresponds to the first site, 2 to the second, etc.) and bits 33 to 64 to spin down, for site 1 to 32. Each bit is set to 1 if it is occupied by an electron of the given spin. In principle, the number of electron flavors and orbitals is free, however, for now it is fixed to 2 and 1.

A list of states should be generated using States(NSites) with NSites being the number of sites in the model.

jED.CDag_sign — Method

CDag_sign(state,i)

Sign when annihilating an electron at position i in state. Returns: -1/1/0 for Uneven/Even permutations and 0 when the state is not occupied.

julia> CDag_sign(SVector{6}(Bool[1,0,1,1,0,0]),5)
-1
julia> CDag_sign(SVector{6}(Bool[1,0,1,1,0,0]),3)
0
julia> CDag_sign(SVector{7}(Bool[1,0,1,0,1,0,0]),4)
1

jED.C_sign — Method

C_sign(state,i)

Sign when creating an electron at position i in state. Returns: -1/1/0 for Uneven/Even permutations and 0 when the state is not occupied.

julia> C_sign(SVector{6}(Bool[1,0,1,1,0,0]),3)
-1
julia> C_sign(SVector{6}(Bool[1,0,1,1,0,0]),2)
0
julia> C_sign(SVector{6}(Bool[1,0,1,1,0,0]),4)

jED.N_el — Method

N_el(s::AbstractVector)::Int

Total electron number of state.

jED.S — Method

S(s::AbstractVector)::Int

Total spin of state. #TODO: only implemented for flavor=2

jED.overlap_cdagger_c — Method

overlap_cdagger_c(left::Fockstate, right::Fockstate, i::Int, j::Int)

Calculate ⟨left| c^†i cj |right⟩, i.e. i is the index for the creation operator and j the index for the annihilation operator. Internally, first check, that both states have exactly two or no difference. In both cases the product of C_sign and CDag_sign is returned, otherwise 0.

Returns: -1/0/1